@@ -1,4 +1,4 @@
-name: nector_grocery
+name: grocery_uikit
 description: A new Flutter project.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: '>=3.0.6 <4.0.0'
 
 dependencies:
   flutter:
@@ -15,19 +15,19 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  google_fonts: ^2.1.1
-  device_preview: ^0.7.1
-  random_color: ^1.0.6-nullsafety
-  flutter_rating_bar: ^4.0.0
-  another_flushbar: ^1.10.28
-  email_validator: ^2.0.1
+  cupertino_icons: ^1.0.6
+  google_fonts: ^6.1.0
+  device_preview: ^1.1.0
+  random_color: ^1.0.5
+  flutter_rating_bar: ^4.0.1
+  another_flushbar: ^1.12.30
+  email_validator: ^2.1.17
   flutter_swiper_null_safety: ^1.0.2
-  carousel_slider: ^4.0.0
-  carousel_pro_nullsafety: ^1.0.2
-  share_plus: ^4.0.3
-  path_provider: ^2.0.9
-  permission_handler: ^9.2.0
+  carousel_slider: ^4.2.1
+  # carousel_pro_nullsafety: ^2.0.0
+  share_plus: ^7.2.1
+  path_provider: ^2.1.2
+  permission_handler: ^11.2.0
 
 dev_dependencies:
   flutter_test:
@@ -38,7 +38,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
