@@ -1,19 +1,15 @@
-
 import 'package:flutter/material.dart';
-import 'package:nector_grocery/constance/routepath.dart' as routepath;
-import 'package:nector_grocery/login_page/splash_screen.dart';
+import 'package:grocery_uikit/constance/routepath.dart' as routepath;
+import 'package:grocery_uikit/login_page/splash_screen.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-
- runApp(const MyApp());
- //runApp(DevicePreview(enabled: !kReleaseMode, builder: (context) => const MyApp()));
+  runApp(const MyApp());
+  //runApp(DevicePreview(enabled: !kReleaseMode, builder: (context) => const MyApp()));
 }
 
-
-
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   // This widget is the root of your application.
   @override
